Pushing to package the program for the team.

+++ include/utils/types.h
@@ -5,5 +5,6 @@
 #ifndef HIGHPERFORMANCELINUXSENSORS_TYPES_H
 #define HIGHPERFORMANCELINUXSENSORS_TYPES_H
 typedef int pid_t;
-typedef int syscall_t;
+typedef unsigned long int syscall_t;
+//typedef int syscall_t;
 #endif //HIGHPERFORMANCELINUXSENSORS_TYPES_H
